fixed error handling

+++ unsplash-scraper.js
@@ -18,15 +18,15 @@ function getUnsplashImage(host, hostURL, dlDir, dlPage, tryNum) {
 		if (dlPage.indexOf("download") > -1 && resp) {
                 	var fileUrl = resp.request.uri.href.split('?').shift();
                 	var fileName = url.parse(fileUrl).pathname.split('/').pop().split('?').shift();
+			if (fileName.indexOf(".jpg") == -1) fileName = fileName + ".jpg";
 			console.log("Started: " + fileUrl);
 
                 	//in curl we have to escape '&' from fileUrl
 			if (!fs.existsSync(dlDir+fileName)) {
-				if (fileName.indexOf(".jpg") == -1) fileName = fileName + ".jpg";
                 		var curl =  'curl ' + fileUrl.replace(/&/g,'\\&') +' -o ' + dlDir+fileName + ' --create-dirs';
                 		var child = exec(curl, function(err, stdout, stderr) {
                    	 		if (err){ 
-						fs.unlinkSync(dlDir+fileName);
+						if (fs.existsSync(dlDir+fileName)) fs.unlinkSync(dlDir+fileName);
 						console.log("Failed: " + fileUrl + " Try: " + tryNum);
 						if (tryNum <= 5) {
 							getUnsplashImage(host, hostURL, dlDir, dlPage, tryNum+1)
